requirements.txt
agentmail>=0.0.60
python-dotenv>=1.1.0
Flask>=3.0.0
requests>=2.31.0
prometheus-client>=0.19.0
pydantic>=2.11.0
apscheduler>=3.10.4
gunicorn>=21.2.0
